+++ ReleaseNotes.md
@@ -2,6 +2,8 @@
 * All UI related texts stored in resources files. Resources for default and pl-PL culture created.
 * Validation messages (RequiredAttribute) localized. 
 * Pizza.Utils assembly created.
+* ValueInjection namespace as top-level in Framework.
+* Useless ObjectInjectionExtensions.InjectFrom removed.
 
 ## New in 0.2.2 (Released 2016-03-21)
 * GridServiceBase renamed to CrudServiceBase